src/cloud_functions/gold/requirements.txt
duckdb==1.4.3
functions-framework==3.10.0
google-cloud-storage==3.7.0
pandas==2.3.3